% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_py_engine.R
\name{copy_python_code}
\alias{copy_python_code}
\title{Copy sklearn codes into the project root folder}
\usage{
copy_python_code(proj_root_path = NULL)
}
\arguments{
\item{proj_root_path}{}
}
\value{
a newly copied folder "rpwf" under the provided project path
}
\description{
these python codes are needed to work with the database. After copying the
codes with \code{\link[=copy_python_code]{copy_python_code()}}, navigate to the "rpwf" folder and run
\verb{python -m pip install -e .} to install the codes as a local python package.
}
\examples{
copy_python_code(here::here())
}
