% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_schema.R
\name{DbCon}
\alias{DbCon}
\title{R6 object store the connection and path to the db}
\description{
Create the "rpwfDb" folder in the provided root path and create a db if
needed. Initialize with \verb{db = DbCon$new(<db_name>, here::here())}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{db_name}}{name of the database}

\item{\code{proj_root_path}}{root path of the project}

\item{\code{db_path}}{path to the \code{db_path}}

\item{\code{con}}{a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DbCon-new}{\code{DbCon$new()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DbCon-new"></a>}}
\if{latex}{\out{\hypertarget{method-DbCon-new}{}}}
\subsection{Method \code{new()}}{
Create an \code{DbCon} object. Should be a singleton. \code{self$con} returns the
\code{\link[DBI:dbConnect]{DBI::dbConnect()}} and \code{self$proj_root_path} returns the \code{proj_root_path}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DbCon$new(db_name, proj_root_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db_name}}{name of the database}

\item{\code{proj_root_path}}{root path of the project}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{DbCon} object.
}
}
}
