% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_simulate.R
\name{dummy_con_}
\alias{dummy_con_}
\title{Generate Temporary Database and Connection}
\usage{
dummy_con_(tmp_dir)
}
\arguments{
\item{tmp_dir}{tmp_dir path.}
}
\value{
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object.
}
\description{
Meant to be called in a \code{withr::local_tempdir()} environment of a test.
}
\keyword{internal}
