% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_tools.R
\name{id_col_switch_}
\alias{id_col_switch_}
\title{Switch function to find what the corresponding id column of tables}
\usage{
id_col_switch_(tbl)
}
\arguments{
\item{tbl}{table names, i.e., "cost_tbl", "df_tbl", etc. Throw error if
tries to return id of "wflow_tbl"}
}
\value{
a character string of id column for respective table.
}
\description{
Not meant to be called manually.
}
\examples{
id_col_switch_("cost_tbl")
id_col_switch_("wflow_tbl") # error
}
