% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_main_R6_classes.R
\name{RGrid}
\alias{RGrid}
\title{Internal R6 object that process the hyper param grid generated in R}
\description{
A R6 object that manage the export of metadata and parquet file of the
hyper param grid generated in R
}
\section{Super class}{
\code{\link[rpwf:Base]{rpwf::Base}} -> \code{RGrid}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RGrid-new}{\code{RGrid$new()}}
\item \href{#method-RGrid-clone}{\code{RGrid$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rpwf" data-topic="Base" data-id="create_folder"><a href='../../rpwf/html/Base.html#method-Base-create_folder'><code>rpwf::Base$create_folder()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpwf" data-topic="Base" data-id="exec_query"><a href='../../rpwf/html/Base.html#method-Base-exec_query'><code>rpwf::Base$exec_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpwf" data-topic="Base" data-id="export"><a href='../../rpwf/html/Base.html#method-Base-export'><code>rpwf::Base$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpwf" data-topic="Base" data-id="export_db"><a href='../../rpwf/html/Base.html#method-Base-export_db'><code>rpwf::Base$export_db()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpwf" data-topic="Base" data-id="export_parquet"><a href='../../rpwf/html/Base.html#method-Base-export_parquet'><code>rpwf::Base$export_parquet()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpwf" data-topic="Base" data-id="query_path"><a href='../../rpwf/html/Base.html#method-Base-query_path'><code>rpwf::Base$query_path()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpwf" data-topic="Base" data-id="set_db_folder"><a href='../../rpwf/html/Base.html#method-Base-set_db_folder'><code>rpwf::Base$set_db_folder()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpwf" data-topic="Base" data-id="set_export_query"><a href='../../rpwf/html/Base.html#method-Base-set_export_query'><code>rpwf::Base$set_export_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpwf" data-topic="Base" data-id="set_hash"><a href='../../rpwf/html/Base.html#method-Base-set_hash'><code>rpwf::Base$set_hash()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpwf" data-topic="Base" data-id="set_path"><a href='../../rpwf/html/Base.html#method-Base-set_path'><code>rpwf::Base$set_path()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rpwf" data-topic="Base" data-id="set_query_results"><a href='../../rpwf/html/Base.html#method-Base-set_query_results'><code>rpwf::Base$set_query_results()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RGrid-new"></a>}}
\if{latex}{\out{\hypertarget{method-RGrid-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of the RGrid class. These are grids made by
functions such as \code{dials::grid_latin_hypercube} and \code{dials::grid_regular}.
See \code{?rprw::Base} for details of the attributes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RGrid$new(grid_obj, con, proj_root_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{grid_obj}}{(\code{rpwf::rpwf_grid_gen()})\cr
\code{\link[=rpwf_grid_gen]{rpwf_grid_gen()}} performs necessary clean ups before the grid can
be used in python}

\item{\code{con}}{(\code{DBI::dbConnect()})\cr
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object, created by \code{\link[=rpwf_db_con]{rpwf_db_con()}}}

\item{\code{proj_root_path}}{(\code{character()})\cr
input the project root path.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RGrid-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RGrid-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RGrid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
