% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_export.R
\name{rpwf_add_all}
\alias{rpwf_add_all}
\title{Wrapper to generate the object to be exported to the database}
\usage{
rpwf_add_all(
  wflow_obj,
  db_con,
  .grid_fun = NULL,
  ...,
  range = c(1L, 5000L),
  seed = 1234L
)
}
\arguments{
\item{wflow_obj}{object created by the \code{\link[=rpwf_workflow_set]{rpwf_workflow_set()}} function}

\item{db_con}{a \link{DbCon} object}

\item{.grid_fun}{\code{{dials}} functions e.g., \code{grid_latin_hypercube},
\code{grid_max_entropy}, etc.,}

\item{...}{additional arguments for the \code{.grid_fun} functions}

\item{range}{range of seed to sample from}

\item{seed}{random seed to sample the random_state}
}
\value{
tibble with the columns necessary for exporting to db
}
\description{
Wrapper to generate the object to be exported to the database
}
