% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_export.R
\name{rpwf_add_desc}
\alias{rpwf_add_desc}
\title{Adds a short description of each workflow to the workflow_set with \code{\link[dplyr:mutate]{dplyr::mutate()}}}
\usage{
rpwf_add_desc(obj)
}
\arguments{
\item{obj}{object generated by \code{\link[=rpwf_add_model_info]{rpwf_add_model_info()}}}
}
\value{
tibble that contains the additional column "wflow_desc"
}
\description{
Adds a short description of each workflow to the workflow_set with \code{\link[dplyr:mutate]{dplyr::mutate()}}
}
