% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_export.R
\name{rpwf_add_desc_}
\alias{rpwf_add_desc_}
\title{Adds a Short Description of Each Workflow to the \code{workflow_set}}
\usage{
rpwf_add_desc_(obj)
}
\arguments{
\item{obj}{object generated by \code{\link[=rpwf_add_model_info_]{rpwf_add_model_info_()}}.}
}
\value{
tibble that contains the additional column "wflow_desc".
}
\description{
Adds a Short Description of Each Workflow to the \code{workflow_set}
}
\keyword{internal}
