% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_export.R
\name{rpwf_add_dfs}
\alias{rpwf_add_dfs}
\title{Process the train/test data}
\usage{
rpwf_add_dfs(obj, db_con, seed)
}
\arguments{
\item{obj}{obj generated by \code{\link[=rpwf_add_grids]{rpwf_add_grids()}}}

\item{db_con}{a \link{DbCon} object}

\item{seed}{random seed}
}
\value{
tibble with the added column "df_id"
}
\description{
Add the \link{TrainDf} R6 objects using provided recipes. Similar to add_grids,
this object does the heavy lifting of generating the path, parquet, and
updating the database.
}
\details{
For each recipe, initialize a new \link{TrainDf}, call \code{self$export()}, then return
the object. This make sure the same object called twice will just fetch
the result from the previous call.
}
