% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_main_R6_classes.R
\name{rpwf_add_fixed_params}
\alias{rpwf_add_fixed_params}
\title{Internal function to add fixed parameters to the \code{{dials}} grid}
\usage{
rpwf_add_fixed_params(model, fixed_params, r_grid)
}
\arguments{
\item{model}{a model spec object defined by \code{{parsnip}}}

\item{fixed_params}{nested named list of fixed params available for this
\code{{parsnip}} model}

\item{r_grid}{the grid generated by \code{{dials}}}
}
\value{
a r_grid with a column of specified fixed parameter
}
\description{
Add fixed parameters, e.g., \code{trees = 150} to overcome default fixed
parameters in sklearn. Names should be provided as {parsnip} hyper param names
because this will be called before renaming. Arguments are provided internally
in \code{rpwf_grid_gen()}. Not meant to be called manually
}
\examples{
#Input should be of the form of
list(xgboost = list(trees = 100))
}
