% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_export.R
\name{rpwf_add_grid_param}
\alias{rpwf_add_grid_param}
\title{Add Relevant Parameters to the \verb{dials::grid_<functions>}}
\usage{
rpwf_add_grid_param(obj, .grid_fun = NULL, seed, ...)
}
\arguments{
\item{obj}{object generated by \code{\link[=rpwf_add_model_info]{rpwf_add_model_info()}}.}

\item{.grid_fun}{a \verb{dials::grid_<functions>}, e.g., \code{grid_random()},
\code{grid_latin_hypercube()}. Default \code{NULL} assumes that no grid search is
performed and default params in sklearn are used. See details for custom grids.}

\item{seed}{random seed.}

\item{...}{additional arguments for the \code{.grid_fun} functions.}
}
\value{
tibble with the additional column "grids".
}
\description{
Add Relevant Parameters to the \verb{dials::grid_<functions>}
}
\details{
Custom grids can be any functions that accept at least an argument "x", as
the functions in {dials} do. So for example, this function would work:

\code{function(x){invisible(x); return(mtcars)}}
}
\keyword{internal}
