% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_export.R
\name{rpwf_add_grids}
\alias{rpwf_add_grids}
\title{Add the RGrid R6 objects using the provided grids.}
\usage{
rpwf_add_grids(obj, db_con)
}
\arguments{
\item{obj}{obj generated by \link{rpwf_add_grid_param}}

\item{db_con}{a \link{DbCon} object}
}
\value{
tibble with the added "grid_id" column
}
\description{
This object does the heavy lifting of generating the path,
parquet and updating the database of the provided R grids
}
\details{
For each grid, initialize a new \link{RGrid}, call \code{self$export()}, then return
the object. This make sure the same object called twice will just fetch
the result from the previous call. Hence, can't be run in parallel

Grab the hashes from the returned \link{RGrid} objects. at this point, we made
sure that 1) the db is updated, 2) file is exported, 3) file exists. Use the
hashes for getting the grid id
}
