% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_export.R
\name{rpwf_add_model_info_}
\alias{rpwf_add_model_info_}
\title{Add \code{py_base_learner}, \code{py_base_learner_args}, \code{model_mode} to the \code{workflow_set}}
\usage{
rpwf_add_model_info_(obj, con)
}
\arguments{
\item{obj}{output of \code{\link[=rpwf_workflow_set]{rpwf_workflow_set()}}.}

\item{con}{(\code{DBI::dbConnect()})\cr
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object, created by \link{DbCon}. Access with \code{DbCon$new()$con}.}
}
\value{
tibble that contains 3 additional columns "py_base_learner",
"py_base_learner_args", "model_mode".
}
\description{
Add \code{py_base_learner}, \code{py_base_learner_args}, \code{model_mode} to the \code{workflow_set}
}
\keyword{internal}
