% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_workflow_export.R
\name{rpwf_add_model_type_}
\alias{rpwf_add_model_type_}
\title{Query the \code{model_type_id} for the Requested Model}
\usage{
rpwf_add_model_type_(obj, con)
}
\arguments{
\item{obj}{an object generated by \code{\link[=rpwf_add_model_info_]{rpwf_add_model_info_()}}.}

\item{con}{(\code{DBI::dbConnect()})\cr
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object, created by \link{DbCon}, access with the \verb{$con} method.
See examples.}
}
\value{
tibble with the "model_type_id" column added.
}
\description{
Query the \code{model_type_id} for the Requested Model
}
\keyword{internal}
