% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_rpwf_db_schema.R
\name{rpwf_add_py_model}
\alias{rpwf_add_py_model}
\title{Add scikit-learn Model Definitions to Database}
\usage{
rpwf_add_py_model(
  db_con,
  py_module,
  py_base_learner,
  r_engine,
  hyper_par_rename,
  model_mode
)
}
\arguments{
\item{db_con}{an \code{\link[=rpwf_connect_db]{rpwf_connect_db()}} object.}

\item{py_module}{the module in scikit-learn, i.e., \code{"sklearn.ensemble"}.}

\item{py_base_learner}{the base learner in scikit-learn, i.e.,
\code{"RandomForestClassifier"}.}

\item{r_engine}{the engine in parsnip, i.e., \code{"ranger"} or \code{"rpart"}.}

\item{hyper_par_rename}{a named list of equivalent hyper parameters, i.e.,
\code{list(cost_complexity = "ccp_alpha")}.}

\item{model_mode}{\code{"classification"} or \code{"regression"}.}
}
\description{
This function adds or updates scikit-learn model to the database. Use the
module (i.e., "xgboost"), the base learner in (i.e., "XGBClassifier"),
the corresponding \code{{parsnip}} engine (i.e., "xgboost"), the equivalent hyper
parameter names (i.e., "mtry" in \code{{parsnip}} is "colsample_bytree"), and
model mode (i.e., "classification").
}
\examples{
# Generate dummy database
board <- pins::board_temp()
tmp_dir <- tempdir()
db_con <- rpwf_connect_db(paste(tmp_dir, "db.SQLite", sep = "/"), board)
DBI::dbListTables(db_con$con)
DBI::dbGetQuery(db_con$con, "SELECT * FROM model_type_tbl") # before adding
rpwf_add_py_model(
  db_con,
  "sklearn.ensemble",
  "RandomForestClassifier",
  "rpart",
  list(
    cost_complexity = "ccp_alpha",
    tree_depth = "max_depth",
    min_n = "min_samples_split"
  ),
  "classification"
)
DBI::dbGetQuery(db_con$con, "SELECT * FROM model_type_tbl") # after adding
}
