% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_export.R
\name{rpwf_add_random_state}
\alias{rpwf_add_random_state}
\title{Add the random state seeds for python random_state}
\usage{
rpwf_add_random_state(obj, range, seed)
}
\arguments{
\item{obj}{an object generated by \link{rpwf_add_model_type}}

\item{range}{range of seed to sample from}

\item{seed}{random seed to sample the random_state}
}
\value{
tibble with "random_state" column added
}
\description{
Add the random state seeds for python random_state
}
