% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_rpwf_db_tools.R
\name{rpwf_avail_models}
\alias{rpwf_avail_models}
\title{Show Available Models}
\usage{
rpwf_avail_models(db_con)
}
\arguments{
\item{db_con}{an \code{\link[=rpwf_connect_db]{rpwf_connect_db()}} object.}
}
\value{
a data.frame of models available in the database.
}
\description{
Models have to be pre-defined and added to the database. Some models are
already added. Other models can be added with \code{\link[=rpwf_add_py_model]{rpwf_add_py_model()}}. This
functions shows the models currently in the database.
}
\examples{
board <- pins::board_temp()
tmp_dir <- tempdir()
db_con <- rpwf_connect_db(paste(tmp_dir, "db.SQLite", sep = "/"), board)
rpwf_avail_models(db_con)
}
