% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_py_engine.R
\name{rpwf_chk_model_avail}
\alias{rpwf_chk_model_avail}
\title{Internal function to check if the specified model is available in the database}
\usage{
rpwf_chk_model_avail(con, py_module, py_base_learner, r_engine)
}
\arguments{
\item{con}{(\code{DBI::dbConnect()})\cr
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object, created by \link{DbCon}}

\item{py_module}{string of what the py module is e.g., xgboost, sklearn.ensemble}

\item{py_base_learner}{string of what the base learner is e.g., XGBClassifier}

\item{r_engine}{string of the r_engine used}
}
\value{
either an error or "Model found in db"
}
\description{
Internal function to check if the specified model is available in the database
}
\examples{
con = rpwf_db_con("db.SQLite", here::here())
rpwf_chk_model_avail(con, "xgboost", "XGBClassifier", "xgboost")
}
