% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_rpwf_only_df_export.R
\name{rpwf_data_set}
\alias{rpwf_data_set}
\title{Juice data from a Recipe}
\usage{
rpwf_data_set(..., db_con)
}
\arguments{
\item{...}{\code{\link[recipes:recipe]{recipes::recipe()}} objects}

\item{db_con}{an \code{\link[=rpwf_connect_db]{rpwf_connect_db()}} object.}
}
\value{
a \code{rpwf_data_set} object
}
\description{
Create an object that holds the juiced prepped data to be exported to Python.
}
\examples{
board <- pins::board_temp()
tmp_dir <- tempdir()
db_con <- rpwf_connect_db(paste(tmp_dir, "db.SQLite", sep = "/"), board)

r <- recipes::recipe(mpg ~ ., data = mtcars)
d <- rpwf_data_set(r, db_con = db_con)
d
}
