% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_schema.R
\name{rpwf_db_con}
\alias{rpwf_db_con}
\title{A wrapper for \code{\link[DBI:dbConnect]{DBI::dbConnect()}} using \code{\link[RSQLite:SQLite]{RSQLite::SQLite()}}}
\usage{
rpwf_db_con(db_name, proj_root_path)
}
\arguments{
\item{db_name}{a string name of the database}

\item{proj_root_path}{a string that indicate the root of the project.
The use of \code{here::here()} within an Rstudio project is recommended.}
}
\value{
A \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object
}
\description{
Creates, if needed, the \code{rpwfDb} folder in the project root path that's
needed for python codes and a connection to the database in that folder.
}
\examples{
con = rpwf_db_con("db.SQLite", here::here())
con
}
