% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_tools.R
\name{rpwf_db_del_entry}
\alias{rpwf_db_del_entry}
\title{Deletes rows from tables other than \code{wflow_tbl}}
\usage{
rpwf_db_del_entry(tbls, id, con)
}
\arguments{
\item{tbls}{vector of character of table names, i.e., "cost_tbl", "df_tbl",
"model_type_tbl", "r_grid_tbl", "r_grid_tbl", "wflow_result_tbl"}

\item{con}{a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object. Created by \code{\link[=rpwf_db_con]{rpwf_db_con()}}}
}
\description{
The deletion of workflows from \code{wflow_tbl} is specified separately to
avoid mistakes.
}
\examples{
con = rpwf_db_con("db.SQLite", here::here())
rpwf_db_del_single_tbl("cost_tbl", 1, con)
}
