% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_tools.R
\name{rpwf_db_del_wflow}
\alias{rpwf_db_del_wflow}
\title{Delete exported workflows}
\usage{
rpwf_db_del_wflow(id, con)
}
\arguments{
\item{id}{numeric vector of workflow ids to be removed.}

\item{con}{a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object. Created by \code{\link[=rpwf_db_con]{rpwf_db_con()}}}
}
\description{
This function specifically just deletes the workflows from the \code{wflow_tbl}
}
\examples{
# Delete workflows with id from 1 to 99 of the database defined by `con`
con = rpwf_db_con("db.SQLite", here::here())
rpwf_db_del_wflow(1:99, con)
}
