% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_schema.R
\name{rpwf_db_init_values_}
\alias{rpwf_db_init_values_}
\title{Add Initial Values to \code{r_grid_tbl} and \code{model_type_tbl}}
\usage{
rpwf_db_init_values_(con)
}
\arguments{
\item{con}{(\code{DBI::dbConnect()})\cr
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object, created by \link{DbCon}, access with the \verb{$con} method.
See examples.}
}
\value{
Used for side effects.
}
\description{
Add some initial values such as the cost functions and the XGBoost model as
defined in R and Python. Won't update duplicated rows. Expand compatibility
by adding values to this function.
}
\examples{
tmp_dir <- withr::local_tempdir()
db_con <- rpwf_connect_db("db.SQLite", tmp_dir)
DBI::dbListTables(db_con$con)
}
\keyword{internal}
