% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_rpwf_viz.R
\name{rpwf_dm}
\alias{rpwf_dm}
\title{Generate a \code{{dm}} Object of the Database}
\usage{
rpwf_dm(con)
}
\arguments{
\item{con}{(\code{DBI::dbConnect()})\cr
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object, created by \link{DbCon}, access with the \verb{$con} method.
See examples.}
}
\value{
\code{{dm}} object of the created schema.
}
\description{
Working with a \code{{dm}} object can make life easier than writing manual queries
}
\examples{
tmp_dir <- withr::local_tempdir()
db_con <- rpwf_connect_db("db.SQLite", tmp_dir)
dm_obj <- rpwf_dm(db_con$con)
dm_obj
# dm_obj |> dm::dm_draw()
}
