% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_export.R
\name{rpwf_export_db}
\alias{rpwf_export_db}
\title{Export the \code{\link[=rpwf_add_all]{rpwf_add_all()}} object into the database and write the parquets}
\usage{
rpwf_export_db(obj, con)
}
\arguments{
\item{obj}{object generated by \code{\link[=rpwf_add_all]{rpwf_add_all()}}}

\item{con}{(\code{DBI::dbConnect()})\cr
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object, created by \link{DbCon}}
}
\value{
number of rows exported
}
\description{
Export the \code{\link[=rpwf_add_all]{rpwf_add_all()}} object into the database and write the parquets
}
