% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_workflow_export.R
\name{rpwf_export_fns_}
\alias{rpwf_export_fns_}
\title{Generate the Export to DB functions}
\usage{
rpwf_export_fns_(required_cols)
}
\arguments{
\item{required_col}{String of required columns.}
}
\value{
a function that add the required columns to the database.
}
\description{
Generate the Export to DB functions
}
