% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_export.R
\name{rpwf_export_grid}
\alias{rpwf_export_grid}
\title{Add the RGrid R6 Objects Using the Provided Grids.}
\usage{
rpwf_export_grid(obj, db_con)
}
\arguments{
\item{obj}{obj generated by \code{\link[=rpwf_add_model_info_]{rpwf_add_model_info_()}}.}

\item{db_con}{a \link{DbCon} object.}
}
\value{
tibble with the added "grid_id" column
}
\description{
This object does the heavy lifting of generating the path,
parquets, and updating the database of the provided R grids. Since the
function is reading from the database and only generating data as needed,
running this function in parallel is not recommended.
}
\details{
For each grid, initialize a new \link{RGrid}, call \code{self$export()}, then return
the object. This make sure the same object called twice will just fetch
the result from the previous call. Hence, can't be run in parallel.

At this point, we made sure that 1) the db is updated, 2) file is exported,
3) file exists. Use the hashes for getting the grid id.
}
