% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_main_R6_classes.R
\name{rpwf_finalize_params}
\alias{rpwf_finalize_params}
\title{Internal function to finalize the parameters that requires train data}
\usage{
rpwf_finalize_params(model, preproc)
}
\arguments{
\item{model}{a model spec object defined by \code{{parsnip}}}

\item{preproc}{a recipe object defined by \code{{recipes}}}
}
\value{
finalized parameter object
}
\description{
mtry is the number of columns in R, whereas in sklearn it's the proportion of
columns (randomly chosen for a base learner). This function creates the
finalized parameter object in R and pass the number of predictors to
\code{\link[=r_grid_gen]{r_grid_gen()}} to convert from mtry count to mtry proportion for sklearn.
}
