% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_main_R6_classes.R
\name{rpwf_finalize_params_}
\alias{rpwf_finalize_params_}
\title{Internal Function to Finalize the Parameters that Requires Train Data}
\usage{
rpwf_finalize_params_(model, preproc)
}
\arguments{
\item{model}{a model spec object defined by \code{{parsnip}}.}

\item{preproc}{a recipe object defined by \code{{recipes}}.}
}
\value{
finalized parameter object.
}
\description{
This function creates the finalized parameter object in R
}
\keyword{internal}
