% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_main_R6_classes.R
\name{rpwf_grid_gen}
\alias{rpwf_grid_gen}
\title{A wrapper around \verb{dials::grid_<functions>} to create a sklearn suitable grid}
\usage{
rpwf_grid_gen(model, preproc, .grid_fun, ...)
}
\arguments{
\item{model}{a model spec object defined by \code{{parsnip}}}

\item{preproc}{a recipe object defined by \code{{recipes}}}

\item{.grid_fun}{\code{{dials}} functions e.g., \code{grid_latin_hypercube},
\code{grid_max_entropy}, etc.,}

\item{...}{additional arguments for the \code{.grid_fun} functions}
}
\value{
a \code{rpwf_grid} object, which is just a modified \code{{dials}} grid but
made suitable for sklearn
}
\description{
This function takes whatever grid functions that are available in \code{{dials}},
converts it to a sklearn suitable format by renaming the hyper parameters as
well as adding fixed parameters to overwrite the default values in sklearn if
provided and perform some cleaning such as converting R \code{mtry} counts to
Python \code{col_by_sample} proportion.
}
