% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_03_rpwf_main_R6_classes.R
\name{rpwf_grid_gen_}
\alias{rpwf_grid_gen_}
\title{A wrapper around \verb{dials::grid_<functions>} to create a sklearn suitable grid}
\usage{
rpwf_grid_gen_(
  model,
  preproc,
  rename_fns,
  .grid_fun = NULL,
  update_params = NULL,
  ...
)
}
\arguments{
\item{model}{a model spec object defined by \code{{parsnip}}.}

\item{preproc}{a recipe object defined by \code{{recipes}}.}

\item{rename_fns}{a function to rename the grid, generated by
\code{rpwf_grid_rename_()}}

\item{.grid_fun}{a \verb{dials::grid_<functions>}, e.g., \code{\link[dials:grid_regular]{dials::grid_random()}},
\code{\link[dials:grid_max_entropy]{dials::grid_latin_hypercube()}}. Default \code{NULL} assumes that no grid search is
performed and sklearn defaults are used.}

\item{...}{additional arguments for the \code{.grid_fun} functions.}
}
\value{
a \code{rpwf_grid} object, which is just a modified \code{{dials}} grid but
made suitable for sklearn.
}
\description{
This function takes whatever grid functions that are available in \code{{dials}},
converts it to a sklearn suitable format by renaming the hyper parameters as
well as adding fixed parameters to overwrite the default values in sklearn if
provided and perform some cleaning such as converting R \code{mtry} counts to
Python \code{col_by_sample} proportion.
}
\keyword{internal}
