% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_main_R6_classes.R
\name{rpwf_grid_rename}
\alias{rpwf_grid_rename}
\title{Rename Hyper Params to sklearn API Suitable Names}
\usage{
rpwf_grid_rename(x)
}
\arguments{
\item{x}{name of the hyper param in R.}
}
\value{
a string of the corresponding hyper param in sklearn.
}
\description{
sklearn uses different names for the same hyper params in \code{{parsnips}}. This
function renames the names generated by \code{dials::grid_functions} into python
suitable names of the hyper parameters that requires tuning. Called by
\code{\link[dplyr:rename]{dplyr::rename_with()}} in \code{\link[=rpwf_grid_gen]{rpwf_grid_gen()}}. Add more to this function
to expand.
}
\examples{
\dontrun{
rpwf_grid_rename("mtry") # "colsample_bytree"
}
}
\keyword{internal}
