% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_main_R6_classes.R
\name{rpwf_grid_rename_}
\alias{rpwf_grid_rename_}
\title{Generate Functions to Rename Hyper Params to sklearn API Suitable Names}
\usage{
rpwf_grid_rename_(json)
}
\arguments{
\item{json}{json that defines how to rename the parameters.}
}
\value{
a function that returns a string of the corresponding hyper param
in sklearn.
}
\description{
sklearn uses different names for the same hyper params in \code{{parsnips}}. This
function create functions to rename the names generated by
\code{dials::grid_functions} into python suitable names of the hyper parameters
that requires tuning. Called by \code{\link[dplyr:rename]{dplyr::rename_with()}} in \code{\link[=rpwf_grid_gen_]{rpwf_grid_gen_()}}.
}
\examples{
rename_fns <- rpwf_grid_rename_(jsonlite::toJSON(list("mtry" = "colsample_bytree")))
rename_fns(c("mtry", "not_added"))
}
\keyword{internal}
