% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_workflow_export.R
\name{rpwf_parquet_id_}
\alias{rpwf_parquet_id_}
\title{Add Grid Id and Df Id to workflow/data set}
\usage{
rpwf_parquet_id_(obj, db_con)
}
\arguments{
\item{obj}{augmented workflow set or data set}

\item{db_con}{a \link{DbCon} object.}
}
\value{
A tibble with df_id and also grid_id added.
}
\description{
Add Grid Id and Df Id to workflow/data set
}
\keyword{internal}
