% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_export.R
\name{rpwf_query}
\alias{rpwf_query}
\title{Internal function for running query to add ids to the workflow_set}
\usage{
rpwf_query(val, query, con)
}
\arguments{
\item{val}{what to use to query, usually the hash of an object}

\item{query}{SQL query}

\item{con}{(\code{DBI::dbConnect()})\cr
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object, created by \code{\link[=rpwf_db_con]{rpwf_db_con()}}}
}
\value{
a vector of query values
}
\description{
This function is a wrapper for \code{\link[DBI:dbGetQuery]{DBI::dbGetQuery()}} that also perform some
cleaning. Should only return one column and <= 1 row for every query. Used in
every function afterwards.
}
