% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_export.R
\name{rpwf_query2}
\alias{rpwf_query2}
\title{Internal function for running query to add ids to the workflow_set}
\usage{
rpwf_query2(val1, val2, query, con)
}
\arguments{
\item{val1}{value vector to search the tables by}

\item{val2}{value vector to search the tables by}

\item{query}{SQL query}

\item{con}{(\code{DBI::dbConnect()})\cr
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object, created by \code{\link[=rpwf_db_con]{rpwf_db_con()}}}
}
\value{
a vector of query values
}
\description{
Internal function for running query to add ids to the workflow_set
}
