% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_workflow_export.R
\name{rpwf_query_}
\alias{rpwf_query_}
\title{Helper Function for Running Query to Add Ids to the \code{workflow_set}}
\usage{
rpwf_query_(query, con, ...)
}
\arguments{
\item{query}{SQL query.}

\item{con}{(\code{DBI::dbConnect()})\cr
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object, created by \link{DbCon}, access with the \verb{$con} method.
See examples.}

\item{...}{value vectors to search the tables by. Accept up to 3 values.}
}
\value{
a vector of query values.
}
\description{
This function is a wrapper for \code{\link[DBI:dbGetQuery]{DBI::dbGetQuery()}} that also perform some
cleaning. Should only return one column and <= 1 row for every query.
}
\keyword{internal}
