% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_04_rpwf_workflow_export.R
\name{rpwf_Rgrid_R6_}
\alias{rpwf_Rgrid_R6_}
\title{Add Rgrid R6 Object}
\usage{
rpwf_Rgrid_R6_(obj, db_con)
}
\arguments{
\item{obj}{a \code{\link[=rpwf_workflow_set]{rpwf_workflow_set()}} object.}

\item{db_con}{a \code{\link[=rpwf_connect_db]{rpwf_connect_db()}} object.}
}
\value{
A tibble with Rgrid objects added.
}
\description{
This function does the heavy lifting of generating the path,
write the parquet, updating the database, for the provided R grids. Since the
function is reading from the database and only generate data as needed,
running this function in parallel is not recommended.
}
\keyword{internal}
