% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_schema.R
\name{rpwf_schema}
\alias{rpwf_schema}
\title{Function contains the db schema}
\usage{
rpwf_schema()
}
\value{
a named list with containing SQL query strings defining a table
}
\description{
Wrapper around the SQL queries that holds the table definitions

Here are the tables available in the database. Print out the raw SQL code
with \verb{rpwf_schema()$<table name>} to see constraints and variable names.
\itemize{
\item \strong{cost_tbl}: contains the cost functions
\item \strong{model_type_tbl}: defines comparable models in R and python
\item \strong{df_tbl}: holds the transformed data (train and test) to pass to python
\item \strong{r_grid_tbl}: holds hyper param grids created in R
\item \strong{wflow_tbl}: defines all the params needed to run a ML model in SKlearn
\item \strong{wflow_result_tbl}: holds results of the workflow ran by python
}
}
\examples{
definitions = rpwf_schema()
names(definitions)
}
