% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_main_R6_classes.R
\name{rpwf_transform_grid_}
\alias{rpwf_transform_grid_}
\title{Transform the Hyper Parameter Grid}
\usage{
rpwf_transform_grid_(grid, rename_fns, n_predictors)
}
\arguments{
\item{grid}{the generated R grid}

\item{rename_fns}{a function to rename the grid, generated by
\code{\link[=rpwf_grid_rename_]{rpwf_grid_rename_()}}}

\item{...}{other parameters needed for transformation}
}
\value{
a transformed and renamed grid
}
\description{
Some hyper parameters between R and Python are transformed differently. This
function is a temporary solution for transformation before a more scalable
solution specific to each model is needed.
}
\keyword{internal}
