% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_export.R
\name{rpwf_wflow_hash_}
\alias{rpwf_wflow_hash_}
\title{Hash the Rows of the Workflow to Check if its Already in the db}
\usage{
rpwf_wflow_hash_(df)
}
\arguments{
\item{df}{object created by \code{\link[=rpwf_add_all]{rpwf_add_all()}}.}
}
\value{
a vector of \link[rlang:hash]{rlang::hash}.
}
\description{
Export the object that has been processed by the other functions into the
database so that python can query the information.
}
\keyword{internal}
