% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_export.R
\name{rpwf_workflow_set}
\alias{rpwf_workflow_set}
\title{Customize version of \code{{workflowsets}}}
\usage{
rpwf_workflow_set(preprocs, models, costs)
}
\arguments{
\item{preprocs}{list or vector of \code{\link[recipes:recipe]{recipes::recipe()}}}

\item{models}{list or vector of model spec. Generated by adding
\code{\link[=set_py_engine]{set_py_engine()}} to a model, e.g. \code{\link[parsnip:boost_tree]{parsnip::boost_tree()}} and
\code{\link[parsnip:set_engine]{parsnip::set_engine()}}}

\item{costs}{list or vector of sklearn cost optimization metrics such as
"neg_log_loss" and "roc_auc", add more by adding rows to the SQL \code{cost_tbl}}
}
\value{
tibble that contains a combination of list of recipes, models,
and costs
}
\description{
Customize version of \code{{workflowsets}}
}
