% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_rpwf_db_schema.R,
%   R/04_rpwf_s3_constructors.R
\name{rpwf_write_board_yaml.pins_board_folder}
\alias{rpwf_write_board_yaml.pins_board_folder}
\alias{rpwf_write_board_yaml}
\title{Print the Board Information as a YAML File}
\usage{
\method{rpwf_write_board_yaml}{pins_board_folder}(board, file)

rpwf_write_board_yaml(board, file)
}
\arguments{
\item{board}{\code{pins::board} object}
}
\value{
named list for YAML
}
\description{
The board information is needed as an input for the pre-made sklearn CLI.
}
\examples{
tmp_dir <- withr::local_tempdir()
board <- pins::board_temp()
rpwf_write_board_yaml(board, paste(tmp_dir, "board.yml", sep = "/"))
list.files(tmp_dir)
}
