% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_rpwf_workflow_export.R
\name{rpwf_write_grid}
\alias{rpwf_write_grid}
\title{Write the Hyper Param Grid Parquet File}
\usage{
rpwf_write_grid(obj)
}
\arguments{
\item{obj}{obj generated by \code{\link[=rpwf_augment]{rpwf_augment()}}.}
}
\value{
Parquet files.
}
\description{
Write the Hyper Param Grid Parquet File
}
\details{
For each grid, initialize a new \link{RGrid}, call \code{self$export()}, then return
the object. This make sure the same object called twice will just fetch
the result from the previous call. Hence, can't be run in parallel.

At this point, we made sure that 1) the db is updated, 2) file is exported,
3) file exists. Use the hashes for getting the grid id.
}
