% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_rpwf_main_R6_classes.R
\name{set_r_grid}
\alias{set_r_grid}
\title{Create a Tune Grid Specific to the Provided Model Spec}
\usage{
set_r_grid(obj, grid_fun, hyper_par_fun = NULL, ...)
}
\arguments{
\item{obj}{a model spec object generated by \code{\link[parsnip:set_engine]{parsnip::set_engine()}} and
\code{\link[parsnip:set_args]{parsnip::set_mode()}}.}

\item{grid_fun}{a \verb{dials::grid_<functions>}, e.g., \code{grid_random()},
\code{grid_latin_hypercube()}.}

\item{hyper_par_fun}{update {dials} functions such as \code{\link[dials:penalty]{dials::penalty()}}
and \code{\link[dials:cost]{dials::cost()}} with a named list. See examples.}

\item{...}{arguments to pass to \code{grid_fun}.}
}
\value{
Add the \code{{dials}} function and associated arguments to the model spec
}
\description{
The hyper parameter grid can either be provided with this function or all at
once later on with \code{\link[=rpwf_augment]{rpwf_augment()}}
}
\examples{
m <- parsnip::boost_tree() |>
  parsnip::set_engine("xgboost") |>
  parsnip::set_mode("classification") |>
  set_r_grid(dials::grid_random, list(trees = dials::trees(range = c(2, 3))), size = 5)
m$grid_fun
m$grid_fun_args
}
