% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_py_related_r_codes.R
\name{update_role_pd_index}
\alias{update_role_pd_index}
\title{Change a Column Role to "pd.index"}
\usage{
update_role_pd_index(obj, index_col)
}
\arguments{
\item{obj}{recipe object}

\item{index_col}{index column}
}
\value{
recipe object with a variable that has the "pd.index" role.
}
\description{
{rpwf} reserves "pd.index" as a special role for pandas index. This function
is a wrapper for \code{\link[recipes:roles]{recipes::update_role()}} where the new role is a "pd.index"
}
\examples{
r <- recipes::recipe(mpg ~ ., data = mtcars) |>
  update_role_pd_index(cyl)
r
}
