% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_sim.R
\name{xgb_model_spec_}
\alias{xgb_model_spec_}
\title{Create a dummy xgb model spec}
\usage{
xgb_model_spec_()
}
\value{
a model spec objection
}
\description{
For testing, not meant to be called, add set_py_engine() to this object
}
