% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_simulate.R
\name{xgb_model_spec_no_tune_}
\alias{xgb_model_spec_no_tune_}
\title{Create a dummy xgb model spec}
\usage{
xgb_model_spec_no_tune_()
}
\value{
a model spec objection
}
\description{
For testing, not meant to be called, have no specified tuning parameters to
test edge case of grid generation. Add set_py_engine() to this object
}
