% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_schema.R
\name{DbCreate}
\alias{DbCreate}
\title{Internal R6 Object that Set and Run a SQL Query}
\description{
A R6 object that provides a shortcut to setting and executing a SQL query
to create new tables in the database. Not meant to be called manually.
}
\examples{

## ------------------------------------------------
## Method `DbCreate$new`
## ------------------------------------------------

db_con = DbCon$new("db.SQLite", tempdir())
db = DbCreate$new(db_con$con, "SELECT * FROM wflow_tbl")
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{con}}{a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object, created by \link{DbCon}.}

\item{\code{query}}{pre-defined SQL query to create a table.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DbCreate-new}{\code{DbCreate$new()}}
\item \href{#method-DbCreate-set_query}{\code{DbCreate$set_query()}}
\item \href{#method-DbCreate-execute}{\code{DbCreate$execute()}}
\item \href{#method-DbCreate-run}{\code{DbCreate$run()}}
\item \href{#method-DbCreate-clone}{\code{DbCreate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DbCreate-new"></a>}}
\if{latex}{\out{\hypertarget{method-DbCreate-new}{}}}
\subsection{Method \code{new()}}{
Create an \code{DbCreate} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DbCreate$new(con, query)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{con}}{a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} connection to the database.}

\item{\code{query}}{a SQL query string.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{DbCreate} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{db_con = DbCon$new("db.SQLite", tempdir())
db = DbCreate$new(db_con$con, "SELECT * FROM wflow_tbl")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DbCreate-set_query"></a>}}
\if{latex}{\out{\hypertarget{method-DbCreate-set_query}{}}}
\subsection{Method \code{set_query()}}{
Change the query
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DbCreate$set_query(query)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{a new SQL query string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DbCreate-execute"></a>}}
\if{latex}{\out{\hypertarget{method-DbCreate-execute}{}}}
\subsection{Method \code{execute()}}{
Run the SQL query
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DbCreate$execute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DbCreate-run"></a>}}
\if{latex}{\out{\hypertarget{method-DbCreate-run}{}}}
\subsection{Method \code{run()}}{
Wrapper around \code{self$set_query()} and \code{self$execute()} that set query and
run the query.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DbCreate$run(new_table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_table}}{a new SQL query string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DbCreate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DbCreate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DbCreate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
