% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_sim.R
\name{dummy_recipe_}
\alias{dummy_recipe_}
\title{Create a dummy recipe object for testing}
\usage{
dummy_recipe_(sim_dat, type = "train")
}
\arguments{
\item{sim_dat}{a \code{\link[=rpwf_sim]{rpwf_sim()}} object}

\item{type}{either \code{"train"} or \code{"test"}}
}
\value{
a \code{\link[recipes:recipe]{recipes::recipe()}} object
}
\description{
For testing, not meant to be called
}
\examples{
test_recipe_train = dummy_recipe_(rpwf_sim)
test_recipe_test = dummy_recipe_(rpwf_sim, type = "test")
}
