% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_export.R
\name{rpwf_add_cost}
\alias{rpwf_add_cost}
\title{Query the cost_id associated with the requested cost function}
\usage{
rpwf_add_cost(obj, con)
}
\arguments{
\item{obj}{object generated by \code{\link[=rpwf_add_dfs]{rpwf_add_dfs()}}}

\item{con}{(\code{DBI::dbConnect()})\cr
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object, created by \link{DbCon}}
}
\value{
tibble with the "cost_id" column added
}
\description{
Query the cost_id associated with the requested cost function
}
