% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_export.R
\name{rpwf_add_model_type}
\alias{rpwf_add_model_type}
\title{Query the model_type id for the requested model}
\usage{
rpwf_add_model_type(obj, con)
}
\arguments{
\item{obj}{an object generated by \code{\link[=rpwf_add_cost]{rpwf_add_cost()}}}

\item{con}{(\code{DBI::dbConnect()})\cr
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object, created by \code{\link[=rpwf_db_con]{rpwf_db_con()}}}
}
\value{
tibble with the "model_type_id" column added
}
\description{
Query the model_type id for the requested model
}
