% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_rpwf_db_tools.R
\name{rpwf_avail_models}
\alias{rpwf_avail_models}
\title{Show Available Models}
\usage{
rpwf_avail_models(con)
}
\arguments{
\item{con}{(\code{DBI::dbConnect()})\cr
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object, created by \link{DbCon}, access with the \verb{$con} method.
See examples.}
}
\value{
a data.frame of models available in the database.
}
\description{
Models have to be pre-defined and added to the database. Some models are
already added. Other models can be added with \code{\link[=rpwf_add_py_model]{rpwf_add_py_model()}}. This
functions shows the models currently in the database.
}
\examples{
tmp_dir <- withr::local_tempdir()
db_con <- rpwf_connect_db("db.SQLite", tmp_dir)
rpwf_avail_models(db_con$con)
}
