% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_py_engine.R
\name{rpwf_chk_model_avail_}
\alias{rpwf_chk_model_avail_}
\title{Internal Function to check if the Specified Model is Available in the Database}
\usage{
rpwf_chk_model_avail_(con, py_module, py_base_learner, r_engine)
}
\arguments{
\item{con}{(\code{DBI::dbConnect()})\cr
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object, created by \link{DbCon}, access with the \verb{$con} method.
See examples.}

\item{py_module}{string of what the py module is e.g., \code{xgboost},
\code{sklearn.ensemble}.}

\item{py_base_learner}{string of what the base learner is e.g.,
\code{XGBClassifier}.}

\item{r_engine}{string of the r_engine used.}
}
\value{
either an error or "Model found in db".
}
\description{
Internal Function to check if the Specified Model is Available in the Database
}
\examples{
tmp_dir <- withr::local_tempdir()
db_con <- rpwf_connect_db("db.SQLite", tmp_dir)
rpwf_db_init_(db_con$con, rpwf_schema()) # Create the database
rpwf_chk_model_avail_(db_con$con, "xgboost", "XGBClassifier", "xgboost")
}
\keyword{internal}
