% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_rpwf_db_schema.R
\name{rpwf_connect_db}
\alias{rpwf_connect_db}
\title{Create a Database and Return an Object that Stores the Connection and Path}
\usage{
rpwf_connect_db(dbname, board, ...)
}
\arguments{
\item{dbname}{path to the database, passed to RSQLite::SQLite().}

\item{board}{a \code{{pins}} board object.}

\item{...}{arguments passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}}
}
\value{
A new \code{DbCon} object.
}
\description{
Create the "rpwfDb" folder in the provided root path and create a db if
needed. Access the connection with \code{object$con}
}
\examples{
board <- pins::board_temp()
tmp_dir <- withr::local_tempdir()
db_con <- rpwf_connect_db(paste(tmp_dir, "db.SQLite", sep = "/"), board)
db_con$con
db_con$board
}
