% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_rpwf_db_schema.R
\name{rpwf_connect_db}
\alias{rpwf_connect_db}
\title{Create a Database and Return an Object that Stores the Connection and Path}
\usage{
rpwf_connect_db(db_name, proj_root_path)
}
\arguments{
\item{db_name}{name of the database.}

\item{proj_root_path}{root path of the project.}
}
\value{
A new \code{DbCon} object.
}
\description{
Create the "rpwfDb" folder in the provided root path and create a db if
needed. Access the connection with \code{object$con}
}
\examples{
tmp_dir <- withr::local_tempdir()
db_con <- rpwf_connect_db("db.SQLite", tmp_dir)
db_con$con
db_con$proj_root_path
}
