% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_py_related_r_codes.R
\name{rpwf_cp_py_codes}
\alias{rpwf_cp_py_codes}
\title{Copy sklearn Codes into the Project Root Folder}
\usage{
rpwf_cp_py_codes(board, overwrite = FALSE)
}
\arguments{
\item{board}{a \code{{pins}} board object.}

\item{overwrite}{overwriting the copied python codes or not. Default to \code{FALSE}.}
}
\value{
a newly copied folder called "rpwf" under the provided project path.
}
\description{
These python codes are needed to work with the database. After copying the
codes with \code{rpwf_cp_py_codes()}, navigate to the "rpwf" folder and run
\verb{python -m pip install -e .} to install the codes as a local python package.
Remove the python codes with \verb{pip uninstall local-rpwf}.
}
\examples{
tmp_dir <- withr::local_tempdir()
rpwf_cp_py_codes(tmp_dir)
list.files(paste0(tmp_dir, "/rpwf"), recursive = TRUE)
}
