% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_py_related_r_codes.R
\name{rpwf_cp_py_codes}
\alias{rpwf_cp_py_codes}
\title{Copy sklearn Codes into the Project Root Folder}
\usage{
rpwf_cp_py_codes(proj_root_path, overwrite = FALSE)
}
\arguments{
\item{proj_root_path}{root path of the project, generate with \link{DbCon} and
assess with DbCon$new()$proj_root_path.}

\item{overwrite}{overwriting the copied python codes or not. Default to FALSE}
}
\value{
a newly copied folder called "rpwf" under the provided project path.
}
\description{
These python codes are needed to work with the database. After copying the
codes with \code{rpwf_cp_py_codes()}, navigate to the "rpwf" folder and run
\verb{python -m pip install -e .} to install the codes as a local python package.
Remove the python codes with \verb{pip uninstall local-rpwf}
}
\examples{
tmp <- tempdir()
rpwf_cp_py_codes(tmp)
list.files(paste0(tmp, "/rpwf"), recursive = TRUE)
}
