% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_py_engine.R
\name{rpwf_cp_py_codes}
\alias{rpwf_cp_py_codes}
\title{Copy sklearn codes into the project root folder}
\usage{
rpwf_cp_py_codes(proj_root_path = NULL)
}
\arguments{
\item{proj_root_path}{}
}
\value{
a newly copied folder "rpwf" under the provided project path
}
\description{
these python codes are needed to work with the database. After copying the
codes with \code{\link[=rpwf_cp_py_codes]{rpwf_cp_py_codes()}}, navigate to the "rpwf" folder and run
\verb{python -m pip install -e .} to install the codes as a local python package.
}
\examples{
rpwf_cp_py_codes(here::here())
}
