% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_rpwf_db_tools.R
\name{rpwf_db_del_entry}
\alias{rpwf_db_del_entry}
\title{Deletes Rows from Tables Other than \code{wflow_tbl}}
\usage{
rpwf_db_del_entry(tbls, id, db_con)
}
\arguments{
\item{tbls}{vector of character of table names, i.e., "df_tbl",
"model_type_tbl", "r_grid_tbl", "r_grid_tbl", "wflow_result_tbl".}

\item{id}{vector of ids to be deleted from a particular table.}

\item{db_con}{an \code{\link[=rpwf_connect_db]{rpwf_connect_db()}} object.}
}
\value{
Called for the side effect.
}
\description{
The deletion of workflows from \code{wflow_tbl} is specified separately to
avoid mistakes.
}
\examples{
board <- pins::board_temp()
tmp_dir <- tempdir()
db_con <- rpwf_connect_db(paste(tmp_dir, "db.SQLite", sep = "/"), board)
# Before deleting
DBI::dbGetQuery(db_con$con, "SELECT * FROM model_type_tbl;")
rpwf_db_del_entry("model_type_tbl", 1, db_con)
# After deleting
DBI::dbGetQuery(db_con$con, "SELECT * FROM model_type_tbl;")
}
