% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_tools.R
\name{rpwf_db_del_entry}
\alias{rpwf_db_del_entry}
\title{Deletes Rows from Tables Other than \code{wflow_tbl}}
\usage{
rpwf_db_del_entry(tbls, id, con)
}
\arguments{
\item{tbls}{vector of character of table names, i.e., "df_tbl",
"model_type_tbl", "r_grid_tbl", "r_grid_tbl", "wflow_result_tbl".}

\item{id}{vector of ids to be deleted from a particular table.}

\item{con}{(\code{DBI::dbConnect()})\cr
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object, created by \link{DbCon}, access with the \verb{$con} method.
See examples.}
}
\value{
Called for the side effect.
}
\description{
The deletion of workflows from \code{wflow_tbl} is specified separately to
avoid mistakes.
}
\examples{
tmp_dir <- withr::local_tempdir()
db_con <- rpwf_connect_db("db.SQLite", tmp_dir)
# Before deleting
DBI::dbGetQuery(db_con$con, "SELECT * FROM model_type_tbl;")
rpwf_db_del_entry("model_type_tbl", 1, db_con$con)
# After deleting
DBI::dbGetQuery(db_con$con, "SELECT * FROM model_type_tbl;")
}
