% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_rpwf_db_tools.R
\name{rpwf_db_del_wflow}
\alias{rpwf_db_del_wflow}
\title{Delete Exported Workflows and Associated Files}
\usage{
rpwf_db_del_wflow(id, db_con)
}
\arguments{
\item{id}{numeric vector of workflow ids to be removed.}

\item{db_con}{(\code{DBI::dbConnect()})\cr
An \code{\link[=rpwf_connect_db]{rpwf_connect_db()}} object.}
}
\value{
Called for the side effect.
}
\description{
This function specifically deletes the workflows from the \code{wflow_tbl}.
}
\examples{
# Delete workflows with id from 1 to 99 of the database defined by `con`
board <- pins::board_temp()
tmp_dir <- withr::local_tempdir()
db_con <- rpwf_connect_db(paste(tmp_dir, "db.SQLite", sep = "/"), board)
rpwf_db_del_wflow(1:99, db_con)
}
