% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_tools.R
\name{rpwf_db_del_wflow}
\alias{rpwf_db_del_wflow}
\title{Delete Exported Workflows}
\usage{
rpwf_db_del_wflow(id, con)
}
\arguments{
\item{id}{numeric vector of workflow ids to be removed.}

\item{con}{(\code{DBI::dbConnect()})\cr
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object, created by \link{DbCon}, access with the \verb{$con} method.
See examples.}
}
\value{
Called for the side effect.
}
\description{
This function specifically just deletes the workflows from the \code{wflow_tbl}.
}
\examples{
# Delete workflows with id from 1 to 99 of the database defined by `con`
tmp_dir <- withr::local_tempdir()
db_con <- rpwf_connect_db("db.SQLite", tmp_dir)
rpwf_db_del_wflow(1:99, db_con$con)
}
