% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_schema.R
\name{rpwf_db_init}
\alias{rpwf_db_init}
\title{Create the Database and Add Initial Values}
\usage{
rpwf_db_init(con, schema = rpwf_schema())
}
\arguments{
\item{con}{(\code{DBI::dbConnect()})\cr
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object, created by \link{DbCon}. Access with \code{DbCon$new()$con}.}

\item{schema}{output of \code{\link[=rpwf_schema]{rpwf_schema()}}.}
}
\value{
Used for side effects.
}
\description{
A wrapper around \link{DbCreate} and \code{\link[=rpwf_db_ini_val]{rpwf_db_ini_val()}}. This function
iteratively runs through the table creation queries in the schema
object and create the tables after creating a new \code{rpwfDb} folder and
database specified by the \code{DbCon$new()} object.
}
\examples{
db_con <- DbCon$new("db.SQLite", tempdir())
rpwf_db_init(db_con$con, rpwf_schema()) # Create the database
DBI::dbListTables(db_con$con)
}
