% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_db_schema.R
\name{rpwf_db_init}
\alias{rpwf_db_init}
\title{Create the database and add initial values}
\usage{
rpwf_db_init(con, schema)
}
\arguments{
\item{con}{(\code{DBI::dbConnect()})\cr
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} object, created by \link{DbCon}}

\item{schema}{a \code{\link[=rpwf_schema]{rpwf_schema()}} object}
}
\description{
A wrapper around \code{\link[=DbCreate$execute]{DbCreate$execute()}} and \code{\link[=rpwf_db_ini_val]{rpwf_db_ini_val()}}.
This function iteratively runs through the table creation queries in the schema
object and create the tables after creating a new \code{rpwfDb} folder and
database specified by the \code{rpwf_db_con()} function.
}
\examples{
db_con = DbCon$new("db.SQLite", here::here())
rpwf_db_init(db_con$con) # Create the database
DBI::dbDisconnect(db_con$con)
}
