% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_rpwf_viz.R
\name{rpwf_dm}
\alias{rpwf_dm}
\title{Generate a \code{{dm}} Object of the Database}
\usage{
rpwf_dm(db_con)
}
\arguments{
\item{db_con}{(\code{DBI::dbConnect()})\cr
An \code{\link[=rpwf_connect_db]{rpwf_connect_db()}} object.}
}
\value{
\code{{dm}} object of the created schema.
}
\description{
Working with a \code{{dm}} object can make life easier than writing manual queries
Also allows the visualization of the database with \code{dm::dm_draw()}.
}
\examples{
board <- pins::board_temp()
tmp_dir <- withr::local_tempdir()
db_con <- rpwf_connect_db(paste(tmp_dir, "db.SQLite", sep = "/"), board)
dm_obj <- rpwf_dm(db_con)
# dm_obj |> dm::dm_draw()
}
