% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_main_R6_classes.R
\name{rpwf_finalize_params_}
\alias{rpwf_finalize_params_}
\title{Internal Function to Finalize the Parameters that Requires Train Data}
\usage{
rpwf_finalize_params_(model, preproc)
}
\arguments{
\item{model}{a model spec object defined by \code{{parsnip}}.}

\item{preproc}{a recipe object defined by \code{{recipes}}.}
}
\value{
finalized parameter object.
}
\description{
\code{mtry} is the number of columns in R, whereas in sklearn it's the proportion
of columns (randomly chosen for a base learner). This function creates the
finalized parameter object in R and pass the number of predictors to
\code{\link[=rpwf_grid_gen_]{rpwf_grid_gen_()}} to convert from \code{mtry} count to \code{col_by_sample} proportion
for sklearn.
}
\keyword{internal}
