% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_rpwf_db_tools.R
\name{rpwf_results}
\alias{rpwf_results}
\title{Get Results from Finished workflows}
\usage{
rpwf_results(db_con, import_csv = TRUE)
}
\arguments{
\item{db_con}{(\code{DBI::dbConnect()})\cr
An \code{\link[=rpwf_connect_db]{rpwf_connect_db()}} object.}

\item{import_csv}{whether to read in the results of the workflow.}
}
\value{
A tibble with the results stored in the fit_results column
}
\description{
Wrapper for an inner join between the \code{wflow_tbl} and \code{wflow_result_tbl} and
invoke \code{\link[utils:read.table]{utils::read.csv()}} to read in the results.
}
