% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_sim.R
\name{rpwf_sim}
\alias{rpwf_sim}
\title{Generate the data needed for testing}
\usage{
rpwf_sim(n_train = 100, n_test = 10, seed = 1234)
}
\arguments{
\item{n_train}{number of train rows}

\item{n_test}{number of test rows}

\item{seed}{random seed}
}
\value{
a list of train and test df and the id and target column names
}
\description{
Generate dummy data for testing of the rpwf package
}
\examples{
dim_dat = rpwf_sim()
}
