% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_rpwf_workflow_export.R
\name{rpwf_tag_recipe}
\alias{rpwf_tag_recipe}
\title{Add a Tag to a Recipe}
\usage{
rpwf_tag_recipe(obj, tag)
}
\arguments{
\item{obj}{An \code{\link[recipes:recipe]{recipes::recipe()}} object.}

\item{tag}{Short description of recipe.}
}
\value{
A tagged \code{\link[recipes:recipe]{recipes::recipe()}} object. Accessible with \code{obj$recipe_tag}.
}
\description{
Complicated workflow sets can become difficult to track. Add a tag to keep
track of unique recipes.
}
\examples{
r <- recipes::recipe(~., data = mtcars)
r <- rpwf_tag_recipe(r, "test recipe")
r$recipe_tag
}
