% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_s3_constructors_and_generics.R,
%   R/01_rpwf_db_schema.R
\name{rpwf_write_board_yaml}
\alias{rpwf_write_board_yaml}
\alias{rpwf_write_board_yaml.pins_board_folder}
\title{Write the Board Information as a YAML File}
\usage{
rpwf_write_board_yaml(board, file)

\method{rpwf_write_board_yaml}{pins_board_folder}(board, file)
}
\arguments{
\item{board}{\code{pins::board} object.}

\item{file}{path to export the YAML file.}
}
\value{
named list for YAML
}
\description{
The board information is needed as an input for the pre-made sklearn CLI.
}
\examples{
board <- pins::board_temp()
tmp_dir <- tempdir()
rpwf_write_board_yaml(board, paste(tmp_dir, "board.yml", sep = "/"))
list.files(tmp_dir)
}
