% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_workflow_export.R
\name{rpwf_write_df}
\alias{rpwf_write_df}
\title{Write the Train/Test Data}
\usage{
rpwf_write_df(obj, db_con, seed = 1234)
}
\arguments{
\item{obj}{obj generated by \code{\link[=rpwf_add_model_info_]{rpwf_add_model_info_()}}.}

\item{db_con}{a \link{DbCon} object.}

\item{seed}{random seed. To control for recipes such as down sampling.}
}
\value{
Parquet files.
}
\description{
Similar to \code{\link[=rpwf_write_grid]{rpwf_write_grid()}}, this function does the heavy lifting of
generating the path, write the parquet, and updating the database for the
recipe of the transformed data. Since the function is reading from the
database and only generating data as needed, running this function in parallel
is not recommended.
}
\details{
For each recipe, initialize a new \link{TrainDf}, call \code{self$export()}, then return
the object. This make sure the same object called twice will just fetch the
result from the previous call.
}
