% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_rpwf_workflow_export.R
\name{rpwf_write_df}
\alias{rpwf_write_df}
\title{Write the Train/Test Data}
\usage{
rpwf_write_df(obj, seed = 1234)
}
\arguments{
\item{obj}{obj generated by \code{\link[=rpwf_augment]{rpwf_augment()}}.}

\item{seed}{random seed. To control for recipes such as down sampling.}
}
\value{
Parquet files.
}
\description{
Write the Train/Test Data
}
\details{
For each recipe, initialize a new \link{TrainDf}, call \code{self$export()}, then return
the object. This make sure the same object called twice will just fetch the
result from the previous call.
}
