% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwf_workflow_export.R
\name{rpwf_write_grid}
\alias{rpwf_write_grid}
\title{Write the Hyper Param Grid Parquet File}
\usage{
rpwf_write_grid(obj, db_con)
}
\arguments{
\item{obj}{obj generated by \code{\link[=rpwf_add_model_info_]{rpwf_add_model_info_()}}.}

\item{db_con}{a \link{DbCon} object.}
}
\value{
Parquet files.
}
\description{
This function does the heavy lifting of generating the path,
write the parquet, updating the database, for the provided R grids. Since the
function is reading from the database and only generate data as needed,
running this function in parallel is not recommended.
}
\details{
For each grid, initialize a new \link{RGrid}, call \code{self$export()}, then return
the object. This make sure the same object called twice will just fetch
the result from the previous call. Hence, can't be run in parallel.

At this point, we made sure that 1) the db is updated, 2) file is exported,
3) file exists. Use the hashes for getting the grid id.
}
