% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_py_related_r_codes.R
\name{set_py_engine}
\alias{set_py_engine}
\title{Define the Python Learner}
\usage{
set_py_engine(obj, py_module, py_base_learner, rpwf_model_tag = NULL, ...)
}
\arguments{
\item{obj}{a model spec object generated by \code{\link[parsnip:set_engine]{parsnip::set_engine()}} and
\code{\link[parsnip:set_args]{parsnip::set_mode()}}.}

\item{py_module}{a character value that select the python module that the
base learner comes from (i.e., \code{xgboost}, \code{sklearn.ensemble}, etc.).}

\item{py_base_learner}{a character value that select the python base learner,
(i.e., XGBClassifier, DecisionTreeClassifier, etc.).}

\item{rpwf_model_tag}{a string to name this model. Useful in keeping track of different
combinations of models.}

\item{...}{arguments passed to the python base learner via named list.
Boolean must be in R format, i.e., \code{TRUE}, \code{FALSE}.}
}
\value{
a model spec object with additional attributes \code{self$rpwf_model_tag},
\code{self$py_base_learner} and \code{self$py_base_learner_args}.
}
\description{
Augment the model spec object generated by \code{{parsnip}} with information
needed to run \code{sklearn} models. Importantly, also passes \code{sklearn} base learner
specific arguments through the \code{args} parameter.
}
\examples{
board <- pins::board_temp()
tmp_dir <- withr::local_tempdir()
db_con <- rpwf_connect_db(paste(tmp_dir, "db.SQLite", sep = "/"), board)
parsnip::boost_tree() |>
  parsnip::set_engine("xgboost") |>
  parsnip::set_mode("classification") |>
  set_py_engine(
    "xgboost",
    "XGBClassifier",
    "my_xgboost_model",
    eval_metric = "logloss",
    use_label_encoder = FALSE,
    verbosity = 0,
    silent = TRUE,
    n_estimators = 100
  )
}
